refactor: split and rename `column_operation.rs` - split out `ColumnType`-related operations to `column_type_operation.rs` - rename the remaining file to `slice_operation.rs`
